% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\docType{package}
\name{bitjson}
\alias{bitjson}
\alias{bitjson-package}
\title{bitjson: Bit JSON}
\description{
De/serialization utility based on JSON bit arrays.
}
\section{Exports}{

  \code{\link{toBitJSON}}
  \code{\link{fromBitJSON}}
  \code{\link{looksLikeBitJSON}}
}

\section{Repository}{

  \url{https://github.com/chiefBiiko/bitjson}
}

