% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress2.R
\name{compressBits}
\alias{compressBits}
\title{Compress a bit vector using chief run-length encoding}
\usage{
compressBits(bits)
}
\arguments{
\item{bits}{Bit vector.}
}
\value{
Integer vector.
}
\description{
Compress a bit vector using chief run-length encoding
}
\keyword{internal}

