% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{decompressBits}
\alias{decompressBits}
\title{Decompress a chief run-length encoded integer vector}
\usage{
decompressBits(encbits)
}
\arguments{
\item{encbits}{Integer vector.}
}
\value{
Bit vector.
}
\description{
Decompress a chief run-length encoded integer vector
}
\keyword{internal}

