% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{fromBitJSON}
\alias{fromBitJSON}
\title{Unserialize an R object encoded as bit JSON}
\usage{
fromBitJSON(x, compressed = TRUE)
}
\arguments{
\item{x}{Bit JSON string or file reference.}

\item{compressed}{Is the bit JSON array compressed via chief run-length
encoding?}
}
\value{
R object.
}
\description{
Unserialize an R object encoded as bit JSON
}
\seealso{
\code{\link{toBitJSON}} \code{\link{looksLikeBitJSON}}
}

