% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{isBitJSON}
\alias{isBitJSON}
\title{Is a JSON string a literal bit array or a bit array compressed via chief run-
length encoding?}
\usage{
isBitJSON(json)
}
\arguments{
\item{json}{JSON string.}
}
\value{
Logical.
}
\description{
TODO: better sequence matching !!!
}
\seealso{
\code{\link{toBitJSON}} \code{\link{fromBitJSON}}
}

