% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDataObject}
\alias{isDataObject}
\title{Is an object a data object?}
\usage{
isDataObject(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Is an object a data object?
}
\details{
Any R object for which \code{is.language} is \code{TRUE}
[calls, expressions, names/symbols] are not considered data objects
in this context.
}
\keyword{internal}

