% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{looksLikeBitJSON}
\alias{looksLikeBitJSON}
\title{Does a JSON string look like a literal bit array or a bit array compressed
via chief run-length encoding?}
\usage{
looksLikeBitJSON(json)
}
\arguments{
\item{json}{JSON string.}
}
\value{
Logical.
}
\description{
Does a JSON string look like a literal bit array or a bit array compressed
via chief run-length encoding?
}
\seealso{
\code{\link{toBitJSON}} \code{\link{fromBitJSON}}
}

