% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serializeToBits}
\alias{serializeToBits}
\title{Serialize any R object to a bit vector.}
\usage{
serializeToBits(x, compress = TRUE)
}
\arguments{
\item{x}{Any R object.}

\item{Logical.}{compress Should the bit vector be compressed via chief
run-length encoding?}
}
\value{
Integer. Compressed bit vector.
}
\description{
Serialize any R object to a bit vector.
}
\keyword{internal}

