% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{toBitJSON}
\alias{toBitJSON}
\title{Serialize an R object to bit JSON}
\usage{
toBitJSON(x, file = NULL, append = FALSE, compress = TRUE, ...)
}
\arguments{
\item{x}{Any R object.}

\item{file}{Character. File name to which to write bit JSON.}

\item{append}{Logical. When writing to a file with \code{cat} overwrite
or append to existing contents?}

\item{compress}{Logical. Compress the return bit array to a chief
run-length encoded integer array?}

\item{...}{Further arguments passed on to \code{jsonlite::toJSON}.}
}
\value{
JSON.
}
\description{
Serialize an R object to bit JSON
}
\seealso{
\code{\link{looksLikeBitJSON}} \code{\link{fromBitJSON}}
}

