% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unSerializeFromBits}
\alias{unSerializeFromBits}
\title{Unserialize an R object from a bit vector.}
\usage{
unSerializeFromBits(x, compressed = TRUE)
}
\arguments{
\item{x}{Integer. Compressed bit vector.}
}
\value{
R object.
}
\description{
Unserialize an R object from a bit vector.
}
\keyword{internal}

