% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{looksLikeBitJSON}
\alias{looksLikeBitJSON}
\title{Does a R object look like a literal bitjson array or a bitjson array 
compressed via chief run-length encoding?}
\usage{
looksLikeBitJSON(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Does a R object look like a literal bitjson array or a bitjson array 
compressed via chief run-length encoding?
}
\seealso{
\code{\link{toBitJSON}} \code{\link{fromBitJSON}}
}

