% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitjson.R
\name{toBitJSON}
\alias{toBitJSON}
\title{Serialize an R object to bit JSON}
\usage{
toBitJSON(x, file = NULL, compress = TRUE)
}
\arguments{
\item{x}{Any R object.}

\item{file}{File name to which to write bit JSON.}

\item{compress}{Compress the return bit array to a chief run-length encoded 
integer array?}
}
\value{
Bit json.
}
\description{
Serialize an R object to bit JSON
}
\seealso{
\code{\link{isBitJSON}} \code{\link{fromBitJSON}}
}

